+++ src/shared/scrapers/GB/ENG/index.js
@@ -1,12 +1,10 @@
 import * as fetch from '../../../lib/fetch/index.js';
 import * as parse from '../../../lib/parse.js';
 
-import gssCodes from './gss-codes.json';
+import { gssCodeMap } from '../_shared.js';
 
 // Set county to this if you only have state data, but this isn't the entire state
 // const UNASSIGNED = '(unassigned)';
-
-// GSS code to ISO code in Wikidata: https://w.wiki/MQ3
 
 const scraper = {
   country: 'iso1:GB',
@@ -14,22 +12,12 @@ const scraper = {
   url: 'https://www.arcgis.com/sharing/rest/content/items/b684319181f94875a6879bbc833ca3a6/data',
   aggregate: 'county',
   type: 'csv',
-  async _gssCodeMap() {
-    const codeMap = {};
-    for (const row of gssCodes) {
-      const { isoCode, gss } = row;
-      codeMap[gss] = isoCode;
-    }
-    // custom code for Bournemouth, Christchurch and Poole
-    codeMap.E06000058 = 'GB-XBCP';
-    return codeMap;
-  },
 
   async scraper() {
     const data = await fetch.csv(this.url);
     const counties = [];
 
-    const codeMap = await this._gssCodeMap();
+    const codeMap = await gssCodeMap();
 
     for (const row of data) {
       const gss = row.GSS_CD;
